airflow-tutorials/tutorial-1/requirements.txt
pandas
numpy
apache-airflow
scikit-learn
slack_sdk
python-dotenv

apache-airflow[amazom]
apache-airflow-providers-amazon
apache-airflow>=2.7.0
apache-airflow-providers-common-sql>=1.3.1
apache-airflow-providers-http
boto3>=1.34.90
botocore>=1.34.90
inflection>=0.5.1
watchtower>=3.0.0,<4
jsonpath_ng>=1.5.3
redshift_connector>=2.0.918
sqlalchemy_redshift>=0.8.6
asgiref>=2.3.0
PyAthena>=3.0.10
jmespath>=0.7.0